+++ app/Http/Controllers/CommentController.php
@@ -14,26 +14,40 @@ use Illuminate\Support\Facades\Log;
 class CommentController extends Controller {
 
     public function store(Request $request, $gameId) {
-        $this->validateComment($request);
-        $user = auth()->user();
+        if(Auth::check()) {
+            $user = auth()->user();
+            $game = Game::findOrFail($gameId);
+            if(in_array($user, $game->users)) {
+                $this->validateComment($request);
 
-        $comment = new Comment([
-            'author' => $user->name,
-            'game_id' => $gameId,
-            'user_id' => $user->id,
-            'text' => $request->text,
-        ]);
+                $comment = new Comment([
+                    'author' => $user->name,
+                    'game_id' => $gameId,
+                    'user_id' => $user->id,
+                    'text' => $request->text,
+                ]);
 
-        $comment->save();
+                $comment->save();
 
-        return redirect()->route('games.show', ['id' => $gameId]);
+                return redirect()->route('games.show', ['id' => $gameId]);
+            }
+        } else {
+            return redirect()->route('login');
+        }
     }
 
     public function delete($id, $gameId) {
-        $comment = Comment::findOrFail($id);
-        $comment->delete();
+        if(Auth::check()) {
+            $comment = Comment::findOrFail($id);
+            $game = Game::findOrFail($gameId);
+            if($comment->user_id == Auth::user()->id || $game->creator->id == Auth::user()->id) {
+                $comment->delete();
 
-        return redirect()->route('games.show', ['id' => $gameId]);
+                return redirect()->route('games.show', ['id' => $gameId]);
+            }
+        } else {
+            return redirect()->route('login');
+        }
     }
 
     public function validateComment(Request $request) {
